+++ com.microej.example.wifi.setup.web/CHANGELOG.md
@@ -1,7 +1,8 @@
-## 1.0.0 (2018-09-26)
+## 1.0.0 (2018-12-28)
 
 Features:
-  - Initial public revision.
+  - Provides a HTML page to list available Wi-Fi using the rest endpoint.
+  - Provides a way to connect to a Wi-Fi using the rest endpoint.
 
 <!--
     Markdown